Moved autosign grains integration tests to its own file
